backend/requirements.txt
autopep8==1.5.4
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.2
django-environ==0.4.5
django-extensions==3.0.9
django-redis==4.12.1
flake8==3.8.4
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
mypy==0.782
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rope==0.17.0
simplegeneric==0.8.1
six==1.15.0
traitlets==5.0.4
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1


djangorestframework==3.12.1
pytest-django==4.1.0
pytest-pythonpath==0.7.3



